% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_module.R
\name{run_app}
\alias{run_app}
\title{Run the Example Shiny App}
\usage{
run_app(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny]{runApp}}.}
}
\value{
Runs the Shiny application.
}
\description{
Launches the example Shiny application included with the package that
demonstrates the histogram modules using the Palmer Penguins dataset.
}
\examples{
\dontrun{
run_app()
}
}
